change dictionary to use int instead of long

+++ src/commonS03DictionaryIntArrayMain/kotlin/lupos/s03resultRepresentation/Value.kt
@@ -1,4 +1,4 @@
 package lupos.s03resultRepresentation
 
 
-typealias Value = Long
+typealias Value = Int
